+++ pkg/agentcompose/capability_proxy.go
@@ -2,8 +2,6 @@ package agentcompose
 
 import (
 	"context"
-	"fmt"
-	"os"
 	"strings"
 
 	"github.com/samber/do/v2"
@@ -31,35 +29,10 @@ func NewCapProxyServer(di do.Injector) (*capproxy.Server, error) {
 }
 
 func (s *Store) ResolveCapabilitySession(ctx context.Context, token string) (capproxy.SessionBinding, error) {
-	token = strings.TrimSpace(token)
-	if token == "" {
-		return capproxy.SessionBinding{}, fmt.Errorf("capability session token is required")
-	}
-	entries, err := os.ReadDir(s.config.SessionRoot)
-	if err != nil {
-		return capproxy.SessionBinding{}, fmt.Errorf("read session root: %w", err)
-	}
-	for _, entry := range entries {
-		if !entry.IsDir() {
-			continue
-		}
-		session, err := s.GetSession(ctx, entry.Name())
-		if err != nil {
-			continue
-		}
-		if sessionCapabilityToken(session) != token {
-			continue
-		}
-		if session.Summary.VMStatus != VMStatusRunning {
-			return capproxy.SessionBinding{}, fmt.Errorf("capability session token is not active")
-		}
-		capsetIDs := sessionCapabilityCapsets(session)
-		if len(capsetIDs) == 0 {
-			return capproxy.SessionBinding{}, fmt.Errorf("session %s has no capability capset", session.Summary.ID)
-		}
-		return capproxy.SessionBinding{SessionID: session.Summary.ID, CapsetIDs: capsetIDs}, nil
-	}
-	return capproxy.SessionBinding{}, fmt.Errorf("capability session token not found")
+	return (httpapi.CapabilitySessionResolver{
+		SessionRoot: s.config.SessionRoot,
+		Store:       s,
+	}).ResolveCapabilitySession(ctx, token)
 }
 
 func sessionCapabilityToken(session *Session) string {
